refactor(combat): move l/s adjustments and harmonization to LocationAndScaleMixin
