+++ src/uniharmony/combat/_ls_mixin.py
@@ -1,5 +1,7 @@
 """Provide LocationAndScaleMixin."""
 
+import math
+
 import numpy as np
 import numpy.typing as npt
 import structlog
@@ -27,6 +29,9 @@ class LocationAndScaleMixin:
         design: npt.NDArray,
         idx_per_site: list[list[int]],
         epsilon: float = 1e-8,
+        delta_epsilon: float = 1e-8,
+        tau_2_epsilon: float = 1e-10,
+        max_iter: int = 1000,
     ) -> tuple[npt.NDArray, list]:
         """Fit L/S model.
 
@@ -41,62 +46,43 @@ class LocationAndScaleMixin:
             e.g., idx_per_site[0] = [0, 5, 10, 15] means samples 0,5,10,15 are from site 0.
         epsilon : float, optional (default 1e-8)
             Small constant to add to variance to avoid division by zero.
-
-        Returns
-        -------
-        gamma_hat : array, shape (n_sites, n_features)
-            Estimated location (mean) shift for each site and each feature.
-            This is how much each site's mean differs from the grand mean.
-        delta_hat : list of array
-            Each array has shape (n_features,) containing estimated scale (variance)
-            for each site and each feature. This is each site's variance.
+        delta_epsilon : float, optional (default 1e-8)
+            Small constant to add to variance to avoid numerical issues in inverse-gamma estimation.
+        tau_2_epsilon : float, optional (default 1e-10)
+            Small constant to add to tau_2 to avoid issues with near-zero variance in gamma_hat.
+        max_iter : int, optional (default 1000)
+            Maximum number of iterations for the solver in full mode.
+            This is only used if empirical_bayes=True and parametric_adjustments=True.
 
         """
         logger.debug("Fitting L/S model")
-        # =====================================================================
-        # STEP 1: Extract site-only design matrix (remove covariate columns)
-        # =====================================================================
+
+        # Extract site-only design matrix (remove covariate columns)
         # The first _n_sites columns of design are site indicator variables (one-hot encoded)
-        # We only want site effects here, not covariate effects
-        # site_design is n_samples x n_sites, where site_design[i,j] = 1 if sample i is from site j
         site_design = design[:, : self._n_sites]
 
-        # =====================================================================
-        # STEP 2: Estimate location parameters (gamma_hat) via OLS
-        # =====================================================================
-        # PURPOSE: Estimate how much each site's mean differs from the grand mean
-        #
-        # MODEL: standardized_data = site_design @ gamma_hat + error
-        #
+        # Estimate location parameters (gamma_hat) via OLS to
+        # check how much each site's mean differs from the grand mean
         # Since data is standardized, we expect gamma_hat to be close to 0
         # Non-zero values indicate site-specific location shifts (batch effects)
-        #
-        # SOLUTION: gamma_hat = (X_site^T @ X_site)^(-1) @ X_site^T @ standardized_data
-        # This gives the OLS estimate of site means on the standardized scale
         gram_site = site_design.T @ site_design
 
         gamma_hat = solve_ordinary_least_squares(gram_site, data, site_design)
 
-        # =====================================================================
-        # STEP 3: Estimate scale parameters (delta_hat) per site
-        # =====================================================================
-        # Estimate each site's variance for each feature
-        #
+        # Estimate scale parameters (delta_hat) per site to
+        # check each site's variance for each feature
         # If sites have different variances, this captures it
-        # delta_hat[j][k] = variance of feature k in site j
-        #
-        # NOTE: We compute this per-site using sample indices, not via OLS
-        # This is because variance is a second-order moment estimated directly
-
+        # We compute this per-site using sample indices, not via OLS as
+        # variance is a second-order moment estimated directly
         delta_hat = []
         for idx, site_idxs in enumerate(idx_per_site):
             if self.mean_only:
-                # [MEAN-ONLY MODE] Assume equal variance across sites
+                # Assume equal variance across sites
                 # Set all variances to 1 (already standardized)
                 # This assumes batch effect is only in location, not scale
                 delta_hat.append(np.repeat(1, data.shape[0]))
             else:
-                # [FULL MODE] Estimate site-specific variances
+                # Estimate site-specific variances
                 # Check minimum samples for variance estimation
                 # With ddof=1, we need at least 2 samples, but more is better
                 n_site_samples = len(site_idxs)
@@ -108,7 +94,6 @@ class LocationAndScaleMixin:
                     delta_hat.append(np.ones(data.shape[0]))
                     continue
                 elif n_site_samples < 16:
-                    # Warn about small sample sizes for variance estimation
                     # Research shows ComBat becomes unstable with <16-32 samples per site
                     logger.warning(
                         f"Site {idx} has only {n_site_samples} samples. "
@@ -134,13 +119,808 @@ class LocationAndScaleMixin:
                 f"Mismatch in site count: expected {self._n_sites} sites, but delta_hat has {len(delta_hat)} entries."
             )
 
-        # =====================================================================
-        # STEP 4: Diagnostic logging of L/S estimates
-        # =====================================================================
-        # Add diagnostic logging in debug mode
         logger.debug("L/S Model estimates:")
         logger.debug(f"  Gamma hat shape: {gamma_hat.shape}")
         logger.debug(f"  Gamma hat range: [{gamma_hat.min():.4f}, {gamma_hat.max():.4f}]")
         for i, d in enumerate(delta_hat):
             logger.debug(f"  Site {i} delta range: [{d.min():.4f}, {d.max():.4f}]")
-        return gamma_hat, delta_hat
+
+        self._find_ls_adjustments(
+            gamma_hat=gamma_hat,
+            delta_hat=delta_hat,
+            data=data,
+            idx_per_site=idx_per_site,
+            delta_epsilon=delta_epsilon,
+            tau_2_epsilon=tau_2_epsilon,
+            max_iter=max_iter,
+        )
+
+    def _find_ls_adjustments(
+        self,
+        gamma_hat: npt.NDArray,
+        delta_hat: list[npt.NDArray],
+        data: npt.NDArray,
+        idx_per_site: list[list[int]],
+        delta_epsilon: float = 1e-8,
+        tau_2_epsilon: float = 1e-10,
+        max_iter: int = 1000,
+    ) -> None:
+        """Find L/S adjustments.
+
+        It can use empirical Bayes to perform either parametric or non-parametric adjustments; or a simple one.
+
+        Parameters
+        ----------
+        gamma_hat : array, shape (n_sites, n_features)
+            Estimated location (mean) shifts for each site and feature from fit_ls_model.
+        delta_hat : list of array
+            Estimated scale (variance) parameters for each site and feature from fit_ls_model.
+            Each array has shape (n_features,).
+        data : array
+            The standardized data if empirical Bayes is used else discarded.
+        idx_per_site : list of list of int
+            List where each element contains sample indices for that site.
+        delta_epsilon : float, optional (default 1e-8)
+            Small constant to add to variance to avoid numerical issues in inverse-gamma estimation.
+        tau_2_epsilon : float, optional (default 1e-10)
+            Small constant to add to tau_2 to avoid issues with near-zero variance in gamma_hat.
+        max_iter : int, optional (default 1000)
+            Maximum number of iterations for the solver in full mode.
+            This is only used if empirical_bayes=True and parametric_adjustments=True.
+
+        """
+        logger.debug("Finding L/S model adjustments")
+
+        # Get the gamma_star and delta_star adjustments to be applied to the data
+        if self.empirical_bayes:
+            logger.debug("Performing empirical Bayes")
+            if self.parametric_adjustments:
+                gamma_bar, tau_2, a_prior, b_prior = self._find_priors(
+                    gamma_hat=gamma_hat,
+                    delta_hat=delta_hat,
+                    delta_epsilon=delta_epsilon,
+                    tau_2_epsilon=tau_2_epsilon,
+                )
+                self._gamma_star, self._delta_star = self._find_parametric_adjustments(
+                    data=data,
+                    idx_per_site=idx_per_site,
+                    gamma_hat=gamma_hat,
+                    delta_hat=delta_hat,
+                    gamma_bar=gamma_bar,
+                    tau_2=tau_2,
+                    a_prior=a_prior,
+                    b_prior=b_prior,
+                    max_iter=max_iter,
+                )
+            else:
+                self._gamma_star, self._delta_star = self._find_non_parametric_adjustments(
+                    data=data,
+                    idx_per_site=idx_per_site,
+                    gamma_hat=gamma_hat,
+                    delta_hat=delta_hat,
+                )
+        else:
+            logger.debug("Finding L/S adjustments without empirical Bayes")
+            self._gamma_star = np.asarray(gamma_hat)
+            self._delta_star = np.asarray(delta_hat)
+
+    def _find_priors(
+        self,
+        gamma_hat: npt.NDArray,
+        delta_hat: list[npt.NDArray],
+        delta_epsilon: float = 1e-8,
+        tau_2_epsilon: float = 1e-10,
+    ) -> tuple[npt.NDArray, npt.NDArray, list[npt.NDArray], list[npt.NDArray]]:
+        """Compute hyperparameters for the prior distributions of batch effects.
+
+        This method estimates the hyperparameters for the Empirical Bayes priors:
+        - Normal prior for location parameters (gamma): N(gamma_bar, tau_2)
+        - Inverse-Gamma prior for scale parameters (delta): IG(a_prior, b_prior)
+
+        These priors are estimated using method of moments from the observed L/S parameters.
+        The priors allow ComBat to "borrow strength" across features, stabilizing estimates
+        especially when sample sizes are small.
+
+        Parameters
+        ----------
+        gamma_hat : array, shape (n_sites, n_features)
+            Estimated location (mean) shifts for each site and feature from _fit_ls_model.
+        delta_hat : list of array
+            Estimated scale (variance) parameters for each site and feature.
+            Each array has shape (n_features,).
+        delta_epsilon : float, optional (default 1e-8)
+            Small constant to add to variance to avoid numerical issues in inverse-gamma estimation.
+        tau_2_epsilon : float, optional (default 1e-10)
+            Small constant to add to tau_2 to avoid issues with near-zero variance in gamma_hat.
+
+        Returns
+        -------
+        gamma_bar : array, shape (n_sites,)
+            Mean of the normal prior for location parameters per site.
+            Represents the expected batch effect across features.
+        tau_2 : array, shape (n_sites,)
+            Variance of the normal prior for location parameters per site.
+            Represents how much batch effects vary across features.
+        a_prior : list of arrays or None
+            Shape parameter of inverse-gamma prior for scale parameters per site.
+            None if mean_only=True (no variance adjustment).
+        b_prior : list of arrays or None
+            Scale parameter of inverse-gamma prior for scale parameters per site.
+            None if mean_only=True (no variance adjustment).
+
+        """
+        logger.debug("Finding priors")
+        # Handle zero/negative variances in delta_hat
+        # Replace them with a small epsilon and log warnings
+        delta_hat_clean = []
+        for idx, site_deltas in enumerate(delta_hat):
+            site_deltas = np.asarray(site_deltas)
+
+            # Check for invalid values (zero, negative, NaN, Inf)
+            invalid_mask = (site_deltas <= 0) | ~np.isfinite(site_deltas)
+
+            if np.any(invalid_mask):
+                n_invalid = np.sum(invalid_mask)
+                logger.warning(
+                    f"Site {idx}: {n_invalid} features have invalid variance values "
+                    f"(<=0, NaN, or Inf). Setting to minimum variance (1e-8). "
+                    "This may indicate constant features or numerical errors."
+                )
+                site_deltas = site_deltas.copy()
+                site_deltas[invalid_mask] = delta_epsilon
+
+            delta_hat_clean.append(site_deltas)
+
+        # Compute mean of gamma_hat across features (axis=1) for each site
+        # This is the expected location effect (gamma_bar) for each site
+        gamma_bar = np.mean(gamma_hat, axis=1)
+
+        # tau_2 represents how much location effects vary across features within each site
+        # High tau_2 = heterogeneous batch effects across features
+        # Low tau_2 = consistent batch effects across features (more pooling possible)
+        tau_2 = np.var(gamma_hat, axis=1, ddof=1)
+
+        # Handle near-zero tau_2 (no variation in batch effects)
+        # This can happen if all features have identical batch effects (rare but possible)
+        # or with very few features
+        tau_2 = handle_near_zero_values(tau_2, epsilon=tau_2_epsilon)
+
+        # Compute inverse-gamma prior parameters for scale (delta)
+        # Estimate a and b using method of moments from the observed delta_hat values
+        # a_prior and b_prior are estimated per feature (not per site) to allow
+        # feature-specific variance pooling
+        if self.mean_only:
+            # No scale adjustment, so no need for inverse-gamma priors
+            a_prior = None
+            b_prior = None
+        else:
+            # Estimate inverse-gamma hyperparameters for each site
+            # _aprior_fn and _bprior_fn implement method of moments estimation:
+            #   a = 1 + mean^2 / variance (shape parameter)
+            #   b = mean * (mean^2 / variance + 1) (scale parameter)
+
+            a_prior = []
+            b_prior = []
+
+            for site_deltas in delta_hat_clean:
+                a_vals, b_vals = self._compute_inverse_gamma_priors(site_deltas)
+
+                a_prior.append(a_vals)
+                b_prior.append(b_vals)
+
+            for idx, (a_vals, b_vals) in enumerate(zip(a_prior, b_prior, strict=True)):
+                invalid_a = (a_vals <= 0) | ~np.isfinite(a_vals)
+                invalid_b = (b_vals <= 0) | ~np.isfinite(b_vals)
+
+                if np.any(invalid_a) or np.any(invalid_b):
+                    n_bad = np.sum(invalid_a | invalid_b)
+                    logger.warning(
+                        f"Site {idx}: {n_bad} features have invalid prior parameters. "
+                        "Clipping to valid range. This may indicate extreme variance values."
+                    )
+                    # Clip to valid ranges for inverse-gamma
+                    a_vals = np.clip(a_vals, 1e-6, 1e6)
+                    b_vals = np.clip(b_vals, 1e-8, 1e8)
+                    a_prior[idx] = a_vals
+                    b_prior[idx] = b_vals
+                    logger.debug("Prior distribution parameters:")
+                    logger.debug(f"  a_prior range: [{a_vals.min():.4f}, {a_vals.max():.4f}]")
+                    logger.debug(f"  b_prior range: [{b_vals.min():.4f}, {b_vals.max():.4f}]")
+
+        logger.debug(f"  Gamma bar (mean location effect): {gamma_bar}")
+        logger.debug(f"  Tau^2 (variance of location effects): {tau_2}")
+        logger.debug(f"  Mean tau^2: {np.mean(tau_2):.6f} (higher = more heterogeneous effects)")
+
+        return gamma_bar, tau_2, a_prior, b_prior
+
+    def _find_parametric_adjustments(
+        self,
+        data: npt.NDArray,
+        idx_per_site: list[list[int]],
+        gamma_hat: npt.NDArray,
+        delta_hat: list[npt.NDArray],
+        gamma_bar: npt.NDArray,
+        tau_2: npt.NDArray,
+        a_prior: list[npt.NDArray],
+        b_prior: list[npt.NDArray],
+        max_iter: int = 1000,
+    ) -> tuple[npt.NDArray, npt.NDArray]:
+        """Compute parametric empirical Bayes site/batch effect parameter estimates.
+
+        This method applies parametric empirical Bayes shrinkage to the L/S model
+        parameters (gamma_hat, delta_hat) to obtain adjusted parameters (gamma_star, delta_star).
+
+        The shrinkage uses the prior distributions estimated in _find_priors:
+        - Normal prior for gamma: shrinks site means toward gamma_bar
+        - Inverse-Gamma prior for delta: shrinks site variances toward common variance
+
+        For mean_only mode, only location (gamma) is adjusted, variance (delta) is fixed at 1.
+        For full mode, both location and variance are adjusted via iterative solver.
+
+        Parameters
+        ----------
+        data : array, shape (n_features, n_samples)
+            Standardized data from _standardize_across_features.
+        idx_per_site : list of list of int
+            List of sample indices for each site.
+        gamma_hat : array, shape (n_sites, n_features)
+            Estimated location parameters from _fit_ls_model.
+        delta_hat : list of array
+            Estimated scale parameters from _fit_ls_model.
+        gamma_bar : array, shape (n_sites,)
+            Mean of normal prior for each site.
+        tau_2 : array, shape (n_sites,)
+            Variance of normal prior for each site.
+        a_prior : list of array
+            Shape parameters of inverse-gamma prior for each site.
+        b_prior : list of array
+            Scale parameters of inverse-gamma prior for each site.
+        max_iter : int, optional (default 1000)
+            Maximum number of iterations for the solver in full mode.
+
+        Returns
+        -------
+        gamma_star : array, shape (n_sites, n_features)
+            Adjusted (shrunken) location parameters for each site and feature.
+        delta_star : array, shape (n_sites, n_features)
+            Adjusted (shrunken) scale parameters for each site and feature.
+
+        Raises
+        ------
+        ValueError
+            If the computer delta_star and gamma_star have incorrect shapes.
+
+        """
+        logger.debug("Finding parametric adjustments")
+
+        gamma_star, delta_star = [], []
+        for i, site_idxs in enumerate(idx_per_site):
+            if self.mean_only:
+                # Fix incorrect parameter passing from original code!
+                # Original passes n=1, delta_star=1 which is wrong
+                # Should use actual sample count and delta_hat[i]
+                n_samples = len(site_idxs)
+
+                # Shrink gamma_hat toward gamma_bar using prior precision
+                gamma_adj = self._postmean(
+                    gamma_hat[i],  # Observed site means
+                    gamma_bar[i],  # Prior mean
+                    n_samples,  # Number of samples in site
+                    1.0,  # Fixed variance (standardized)
+                    tau_2[i],  # Prior variance
+                )
+                gamma_star.append(gamma_adj)
+
+                # Fixed variance = 1 (standardized scale)
+                delta_star.append(np.ones(data.shape[0]))
+            else:
+                # Adjust both location and variance iteratively
+                gamma_adj, delta_adj = self._iteration_solver(
+                    data[:, site_idxs],
+                    gamma_hat[i],
+                    delta_hat[i],
+                    gamma_bar[i],
+                    tau_2[i],
+                    a_prior[i],
+                    b_prior[i],
+                    max_iter=max_iter,
+                )
+                gamma_star.append(gamma_adj)
+                delta_star.append(delta_adj)
+        # Stack lists into arrays with proper shape checking
+        gamma_star_arr = np.asarray(gamma_star)
+        delta_star_arr = np.asarray(delta_star)
+
+        expected_shape = (self._n_sites, data.shape[0])
+        if gamma_star_arr.shape != expected_shape:
+            raise ValueError(f"gamma_star shape {gamma_star_arr.shape} != expected {expected_shape}")
+        if delta_star_arr.shape != expected_shape:
+            raise ValueError(f"delta_star shape {delta_star_arr.shape} != expected {expected_shape}")
+        return gamma_star_arr, delta_star_arr
+
+    def _compute_inverse_gamma_priors(self, delta_hat: npt.NDArray) -> tuple[npt.NDArray, npt.NDArray]:
+        """Calculate both shape (a) and scale (b) parameters for inverse-gamma prior.
+
+        Computes both simultaneously since they share moment calculations.
+        This is ~2x faster than calling separate functions.
+
+        Parameters
+        ----------
+        delta_hat : array, shape (n_features,)
+            Estimated variance parameters for a single site.
+
+        Returns
+        -------
+        a_prior : ndarray, shape (n_features,)
+            Shape parameter: a = (2*s2 + m^2) / s2
+        b_prior : ndarray, shape (n_features,)
+            Scale parameter: b = (m*s2 + m^3) / s2
+
+        """
+        delta_hat = np.asarray(delta_hat, dtype=np.float64)
+
+        # Clean input
+        if np.any(delta_hat <= 0) or not np.all(np.isfinite(delta_hat)):
+            logger.warning("Invalid values in delta_hat, clipping to valid range and replacing non-finite values with median")
+            delta_hat = np.clip(delta_hat, 1e-10, 1e10)
+            delta_hat = np.where(np.isfinite(delta_hat), delta_hat, np.median(delta_hat))
+
+        # Compute moments once
+        m = np.mean(delta_hat)
+        s2 = np.var(delta_hat, ddof=1)
+
+        # Handle near-zero variance
+        if s2 < 1e-10:
+            logger.warning("Variance of delta_hat is near zero, using strong pooling for priors")
+            return (
+                np.full_like(delta_hat, 1e6),  # Large a = strong pooling
+                np.full_like(delta_hat, m),  # b = mean
+            )
+
+        # Compute both parameters
+        m2 = m * m
+        a_prior = (2.0 * s2 + m2) / s2
+        b_prior = m * (s2 + m2) / s2
+
+        a_prior = np.clip(a_prior, 1e-6, 1e8, dtype=np.float32)
+        b_prior = np.clip(b_prior, 1e-8, 1e8, dtype=np.float32)
+        return a_prior, b_prior
+
+    def _postmean(
+        self,
+        gamma_hat: npt.ArrayLike,
+        gamma_bar: npt.ArrayLike,
+        n: int | npt.NDArray,
+        delta_star: float | npt.ArrayLike,
+        tau_2: npt.ArrayLike,
+    ) -> npt.NDArray:
+        """Compute posterior mean (shrunken estimate) for location parameters.
+
+        Formula: gamma_star = (tau_2 * n * gamma_hat + delta * gamma_bar) / (tau_2 * n + delta)
+
+        Parameters
+        ----------
+        gamma_hat : array-like, shape (n_features,)
+            Observed site means from L/S model.
+        gamma_bar : array-like, shape (n_features,)
+            Prior mean (expected batch effect across features).
+        n : int or array-like
+            Sample size for the site.
+        delta_star : float or array-like
+            Estimated variance (sampling precision).
+        tau_2 : array-like, shape (n_features,)
+            Prior variance (how much batch effects vary across features).
+
+        Returns
+        -------
+        gamma_star : ndarray, shape (n_features,)
+            Posterior mean (shrunken estimate) for each feature.
+
+        """
+        # Convert to arrays for broadcasting
+        gamma_hat = np.asarray(gamma_hat)
+        gamma_bar = np.asarray(gamma_bar)
+        tau_2 = np.asarray(tau_2)
+        n = np.asarray(n)
+        delta_star = np.asarray(delta_star)
+
+        # Compute posterior mean using precision-weighted average
+        prior_precision = tau_2 * n
+
+        # Handle numerical issues
+        denominator = prior_precision + delta_star
+        if np.any(denominator <= 0):
+            logger.warning("_postmean: Non-positive denominator, clipping")
+            denominator = np.clip(denominator, 1e-10, None)
+
+        return (prior_precision * gamma_hat + delta_star * gamma_bar) / denominator
+
+    def _postvar(
+        self,
+        sum_2: float | npt.ArrayLike,
+        n: int | npt.ArrayLike,
+        a_prior: npt.ArrayLike,
+        b_prior: npt.ArrayLike,
+    ) -> npt.NDArray:
+        """Compute posterior variance estimate for scale parameters.
+
+        Formula: delta_star = (0.5 * sum_2 + b_prior) / (n/2 + a_prior - 1)
+
+        Parameters
+        ----------
+        sum_2 : float or array-like, shape (n_features,)
+            Sum of squared deviations: sum((x - gamma_star)^2)
+        n : int or array-like
+            Sample size.
+        a_prior : array-like, shape (n_features,)
+            Shape parameter of inverse-gamma prior.
+        b_prior : array-like, shape (n_features,)
+            Scale parameter of inverse-gamma prior.
+
+        Returns
+        -------
+        delta_star : ndarray, shape (n_features,)
+            Posterior variance estimate for each feature.
+
+        """
+        # Convert to arrays for broadcasting
+        sum_2 = np.asarray(sum_2)
+        a_prior = np.asarray(a_prior)
+        b_prior = np.asarray(b_prior)
+        n = np.asarray(n)
+
+        # Posterior update for inverse-gamma
+        numerator = 0.5 * sum_2 + b_prior
+        denominator = n / 2.0 + a_prior - 1.0
+
+        # Handle numerical issues
+        if np.any(denominator <= 0):
+            logger.warning("_postvar: Non-positive denominator, using prior estimate only")
+            with np.errstate(divide="ignore", invalid="ignore"):
+                prior_mean = np.where(a_prior > 1, b_prior / (a_prior - 1), b_prior)
+            return prior_mean
+
+        result = numerator / denominator
+
+        # Ensure positive variance
+        return np.clip(result, 1e-8, None)
+
+    def _iteration_solver(
+        self,
+        data: npt.NDArray,
+        gamma_hat: npt.NDArray,
+        delta_hat: npt.NDArray,
+        gamma_bar: npt.NDArray,
+        tau_2: npt.NDArray,
+        a_prior: npt.NDArray,
+        b_prior: npt.NDArray,
+        convergence: float = 0.0001,
+        max_iter: int = 1000,
+    ) -> tuple[npt.NDArray, npt.NDArray]:
+        """Solve iteratively the posterior mean and variance via Empirical Bayes.
+
+        Parameters
+        ----------
+        data : array, shape (n_features, n_samples_site)
+            Standardized data for a single site.
+        gamma_hat : array, shape (n_features,)
+            Initial location estimates.
+        delta_hat : array, shape (n_features,)
+            Initial scale estimates.
+        gamma_bar : array, shape (n_features,)
+            Prior mean for location.
+        tau_2 : array, shape (n_features,)
+            Prior variance for location.
+        a_prior : array shape (n_features,)
+            Prior shape for scale.
+        b_prior : array, shape (n_features,)
+            Prior scale for scale.
+        convergence : float, optional (default 0.0001)
+            Relative change threshold for convergence.
+        max_iter : int, optional (default 1000)
+            Maximum number of iterations.
+
+        Returns
+        -------
+        gamma_star : ndarray, shape (n_features,)
+            Converged posterior location estimates.
+        delta_star : ndarray, shape (n_features,)
+            Converged posterior scale estimates.
+
+        Raises
+        ------
+        ValueError
+            If NaN values are found in ``data``.
+
+        """
+        gamma_hat_new = None
+        delta_hat_new = None
+        # Handle missing data (NaN) by counting non-NaN samples per feature
+        sample_size = np.sum(~np.isnan(data), axis=1)
+        if np.any(sample_size == 0):
+            raise ValueError("Some features have all NaN values for this site")
+
+        gamma_hat_old = np.asarray(gamma_hat, dtype=np.float64).copy()
+        delta_hat_old = np.asarray(delta_hat, dtype=np.float64).copy()
+
+        # Validate inputs
+        if np.any(delta_hat_old <= 0):
+            logger.warning("_iteration_solver: Initial delta_hat <= 0, clipping")
+            delta_hat_old = np.clip(delta_hat_old, 1e-8, None)
+
+        change = 1.0
+        count = 0
+
+        while change > convergence:
+            # Add iteration limit to void infinite loops in edge cases
+            if count >= max_iter:
+                logger.warning(
+                    f"_iteration_solver: Did not converge after {max_iter} iterations. "
+                    f"Final change: {change:.6f}. Using current estimates."
+                )
+                break
+
+            # E-step: Update gamma given current delta
+            gamma_hat_new = self._postmean(gamma_hat, gamma_bar, sample_size, delta_hat_old, tau_2)
+
+            # M-step: Update delta given new gamma
+            # Use broadcasting instead of explicit ones matrix
+            gamma_expanded = gamma_hat_new[:, np.newaxis]
+
+            with np.errstate(invalid="ignore"):
+                residuals = data - gamma_expanded
+                resid2 = np.square(residuals)
+                sum_2 = np.nansum(resid2, axis=1)
+
+            delta_hat_new = self._postvar(sum_2, sample_size, a_prior, b_prior)
+            # Handle any numerical issues in delta_hat_new
+            delta_hat_new = np.clip(delta_hat_new, 1e-8, None)
+
+            # Convergence check
+            with np.errstate(divide="ignore", invalid="ignore"):
+                gamma_change = np.abs(gamma_hat_new - gamma_hat_old)
+                gamma_rel_change = np.where(np.abs(gamma_hat_old) > 1e-10, gamma_change / np.abs(gamma_hat_old), gamma_change)
+
+                delta_change = np.abs(delta_hat_new - delta_hat_old)
+                delta_rel_change = np.where(np.abs(delta_hat_old) > 1e-10, delta_change / np.abs(delta_hat_old), delta_change)
+
+                change = max(np.max(gamma_rel_change), np.max(delta_rel_change))
+
+            gamma_hat_old = gamma_hat_new.copy()
+            delta_hat_old = delta_hat_new.copy()
+            count += 1
+
+        logger.debug(f"_iteration_solver converged in {count} iterations (change={change:.6f})")
+        # Final validation of outputs
+        gamma_hat_new = np.asarray(gamma_hat_new, dtype=np.float64)
+        delta_hat_new = np.asarray(delta_hat_new, dtype=np.float64)
+        return gamma_hat_new, delta_hat_new
+
+    def _find_non_parametric_adjustments(
+        self,
+        data: npt.NDArray,
+        idx_per_site: list[list[int]],
+        gamma_hat: npt.NDArray,
+        delta_hat: list[npt.NDArray],
+    ) -> tuple[npt.NDArray, npt.NDArray]:
+        """Compute non-parametric empirical Bayes site/batch effect parameter estimates.
+
+        Parameters
+        ----------
+        data : array, shape (n_features, n_samples)
+            Standardized data.
+        idx_per_site : list of list of int
+            List of sample indices for each site.
+        gamma_hat : array, shape (n_sites, n_features)
+            Estimated location parameters.
+        delta_hat : list of array
+            Estimated scale parameters.
+
+        Returns
+        -------
+        gamma_star : array, shape (n_sites, n_features)
+            Adjusted location parameters via non-parametric EB.
+        delta_star : array, shape (n_sites, n_features)
+            Adjusted scale parameters via non-parametric EB.
+
+        Raises
+        ------
+        ValueError
+            If the computer delta_star and gamma_star have incorrect shapes.
+
+        """
+        logger.debug("Finding non-parametric adjustments")
+
+        gamma_star, delta_star = [], []
+
+        for i, site_idxs in enumerate(idx_per_site):
+            # Don't modify delta_hat in place
+            if self.mean_only:
+                site_delta = np.ones(data.shape[0])
+            else:
+                site_delta = np.asarray(delta_hat[i])
+
+            gamma_adj, delta_adj = self._int_eprior(
+                data[:, site_idxs],
+                gamma_hat[i],
+                site_delta,
+            )
+
+            gamma_star.append(gamma_adj)
+            delta_star.append(delta_adj)
+
+        # Convert to arrays with validation
+        gamma_star_arr = np.asarray(gamma_star, dtype=np.float64)
+        delta_star_arr = np.asarray(delta_star, dtype=np.float64)
+
+        expected_shape = (self._n_sites, data.shape[0])
+        if gamma_star_arr.shape != expected_shape:
+            raise ValueError(f"gamma_star shape mismatch: {gamma_star_arr.shape} vs {expected_shape}")
+        if delta_star_arr.shape != expected_shape:
+            raise ValueError(f"delta_star shape mismatch: {delta_star_arr.shape} vs {expected_shape}")
+
+        return gamma_star_arr, delta_star_arr
+
+    def _int_eprior(
+        self,
+        data: npt.NDArray,
+        gamma_hat: npt.NDArray,
+        delta_hat: npt.NDArray,
+    ) -> tuple[npt.NDArray, npt.NDArray]:
+        """Compute non-parametric empirical Bayes adjustments via kernel weighting.
+
+        Parameters
+        ----------
+        data : array, shape (n_features, n_samples_site)
+            Standardized data for a single site.
+        gamma_hat : array, shape (n_features,)
+            Location parameters for each feature.
+        delta_hat : array, shape (n_features,)
+            Scale parameters for each feature.
+
+        Returns
+        -------
+        gamma_star : ndarray, shape (n_features,)
+            Weighted location estimates.
+        delta_star : ndarray, shape (n_features,)
+            Weighted scale estimates.
+
+        """
+        n_features = data.shape[0]
+
+        gamma_hat = np.asarray(gamma_hat, dtype=np.float64)
+        delta_hat = np.asarray(delta_hat, dtype=np.float64)
+
+        # Pre-allocate output arrays
+        gamma_star = np.empty(n_features, dtype=np.float64)
+        delta_star = np.empty(n_features, dtype=np.float64)
+
+        # Precompute constants
+        two_pi = 2.0 * math.pi
+
+        for i in range(n_features):
+            x = data[i, :]
+            n = x.shape[0]
+
+            # Leave-one-out: use all OTHER features' parameters
+            g_other = np.delete(gamma_hat, i)
+            d_other = np.delete(delta_hat, i)
+
+            # Vectorized likelihood computation with broadcasting
+            x_expanded = x[np.newaxis, :]
+            g_expanded = g_other[:, np.newaxis]
+
+            residuals = x_expanded - g_expanded
+            resid2 = np.square(residuals)
+            sum2 = np.sum(resid2, axis=1)
+
+            # Use log-space for numerical stability
+            with np.errstate(divide="ignore", invalid="ignore"):
+                log_lh = -0.5 * n * np.log(two_pi * d_other) - sum2 / (2.0 * d_other)
+                log_lh_max = np.max(log_lh)
+                lh = np.exp(log_lh - log_lh_max)
+
+            # Handle numerical issues
+            if np.any(np.isnan(lh)) or np.any(np.isinf(lh)):
+                logger.warning(
+                    f"Feature {i}: Numerical issues in likelihood computation, replacing NaN/Inf with zeros"
+                    "Please check data for constant features or extreme values."
+                )
+                lh = np.nan_to_num(lh, nan=0.0, posinf=0.0, neginf=0.0)
+
+            lh_sum = np.sum(lh)
+            if lh_sum < 1e-300:
+                logger.warning(f"Feature {i}: All likelihoods near zero, using original estimate")
+                gamma_star[i] = gamma_hat[i]
+                delta_star[i] = delta_hat[i]
+                continue
+
+            # Compute weights and weighted averages for gamma and delta
+            weights = lh / lh_sum
+
+            # Weighted average of gamma and delta using the computed weights
+            gamma_star[i] = np.sum(g_other * weights)
+            delta_star[i] = np.sum(d_other * weights)
+
+        return gamma_star, delta_star
+
+    def harmonize(
+        self,
+        data: npt.NDArray,
+        mean: npt.NDArray,
+        idx_per_site: list[list[int]],
+        epsilon: float = 1e-8,
+    ):
+        """Compute the final harmonized data.
+
+        This method applies the L/S adjustments (gamma_star, delta_star)
+        to remove site effects and then rescales the data back to original scale.
+
+        The harmonization formula for each site j:
+            bayes_data = (data - gamma_star[j]) / sqrt(delta_star[j])
+        Then rescale: bayes_data * sqrt(var_pooled) + mean
+
+        Parameters
+        ----------
+        data : array, shape (n_features, n_samples)
+            Standardized data from _standardize_across_features.
+        mean : array, shape (n_features, n_samples)
+            Standardized mean used during standardization.
+        idx_per_site : list of list of int
+            List of sample indices for each site.
+        epsilon : float, optional (default 1e-8)
+            Small constant to add to delta_star to avoid division by zero.
+
+        Returns
+        -------
+        bayes_data : ndarray, shape (n_features, n_samples)
+            Harmonized data in original scale.
+
+        Raises
+        ------
+        ValueError
+            If gamma_star or delta_star are not fitted or have wrong shapes.
+        RuntimeError
+            If numerical issues are detected during harmonization.
+
+        """
+        logger.debug("Harmonizing data")
+
+        # Validate fitted attributes
+        if not hasattr(self, "_gamma_star") or not hasattr(self, "_delta_star"):
+            raise ValueError("gamma_star and delta_star must be computed before adjustment")
+
+        # Validate shapes
+        expected_shape = (self._n_sites, data.shape[0])
+        if self._gamma_star.shape != expected_shape:
+            raise ValueError(f"gamma_star shape {self._gamma_star.shape} != expected {expected_shape}.")
+        if self._delta_star.shape != expected_shape:
+            raise ValueError(f"delta_star shape {self._delta_star.shape} != expected {expected_shape}.")
+        if self._var_pooled.shape[0] != data.shape[0]:
+            raise ValueError(f"var_pooled has {self._var_pooled.shape[0]} features but data has {data.shape[0]} features.")
+
+        # Apply EB adjustments per site
+        # Create output array (copy to avoid modifying input)
+        bayes_data = data.copy()
+        pooled_std = np.sqrt(self._var_pooled)
+        if pooled_std.ndim == 1:
+            pooled_std = pooled_std[:, np.newaxis]
+        for j, site_idxs in enumerate(idx_per_site):
+            if len(site_idxs) == 0:
+                logger.warning(f"Site {j} has no samples")
+                continue
+
+            # Get adjusted parameters for this site
+            site_gamma = self._gamma_star[j, :][:, np.newaxis]  # (n_features, 1)
+            site_delta = np.maximum(self._delta_star[j, :], epsilon)[:, np.newaxis]  # (n_features, 1)
+
+            # Apply harmonization formula
+            bayes_data[:, site_idxs] = (bayes_data[:, site_idxs] - site_gamma) / np.sqrt(site_delta)
+
+        # Rescale to original scale
+        bayes_data = bayes_data * pooled_std + mean
+
+        # Validate output
+        if not np.all(np.isfinite(bayes_data)):
+            raise RuntimeError("Harmonization produced non-finite values")
+        return bayes_data
